move FileApplication to own module
