+++ src/middlewared/middlewared/apps/__init__.py
@@ -1,0 +1,1 @@
+from .file_app import FileApplication
